More minor tweaks and such
git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code@3295 5227b50d-9349-4574-b055-4b0a7525e11c

+++ AWB/WikiFunctions/DatabaseScanner/Scanners.cs
@@ -113,7 +113,7 @@ namespace WikiFunctions.DBScanner
         }
     }
 
-    public enum MoreLessThan : byte { LessThan, MoreThan, EqualTo }
+    public enum MoreLessThan : int { LessThan, MoreThan, EqualTo }
 
     public class CountCharacters : Scan
     {
